Parameters for Classifier Prediction

Collects the function to be used for making predictions and any associated parameters.

Details

The function specified must return either a factor vector of class predictions, or a numeric vector of scores for the second class, according to the levels of the class vector of the input data set, or a data frame which has two columns named class and score.

Constructor

PredictParams(predictor, characteristics = DataFrame(), intermediate = character(0), ...)

Creates a PredictParams object which stores the function which will do the class prediction, if required, and parameters that the function will use. If the training function also makes predictions, this must be set to NULL.

predictParams is a PredictParams object.

show(predictParams): Prints a short summary of what predictParams contains.

Author

Dario Strbenac

Examples

# For prediction by trained object created by DLDA training function.
predictParams <- PredictParams("DLDA")