A function to perform fast or standard Cox proportional hazard model tests.

A function to perform fast or standard Cox proportional hazard model tests.

Usage

colCoxTests(measurements, outcome, option = c("fast", "slow"), ...)

Arguments

measurements

matrix with variables as columns.

outcome

matrix with first column as time and second column as event.

option

Default: "fast". Whether to use the fast or slow method.

...

Not currently used.

Value

CrossValParams object

Examples

data(asthma)
time <- rpois(nrow(measurements), 100)
status <- sample(c(0,1), nrow(measurements), replace = TRUE)
outcome <- cbind(time, status)
output <- colCoxTests(measurements, outcome, "fast")