Various Functions for Evaluating Precision Pathways

These functions tabulate or plot various aspects of precision pathways, such as accuracies and costs.

Usage

calcCostsAndPerformance(precisionPathways, costs = NULL)

# S3 method for class 'PrecisionPathways'
summary(object, weights = c(accuracy = 0.5, cost = 0.5), ...)

# S3 method for class 'PrecisionPathways'
bubblePlot(precisionPathways, pathwayColours = NULL, ...)

# S3 method for class 'PrecisionPathways'
flowchart(
  precisionPathways,
  pathway,
  nodeColours = c(assay = "#86C57C", class1 = "#ACCEE0", class2 = "#F47F72"),
  ...
)

# S3 method for class 'PrecisionPathways'
strataPlot(
  precisionPathways,
  pathway,
  classColours = c(class1 = "#4DAF4A", class2 = "#984EA3"),
  ...
)

Arguments

precisionPathways

A pathway of class PrecisionPathways.

costs

A named vector of assays with the cost of each one.

object

A set of pathways of class PrecisionPathways.

weights

A numeric vector of length two specifying how to weight the predictive accuracy and the cost during ranking. Must sum to 1.

...

Not used but just following the S3 requirement of the generic template.

pathwayColours

A named vector of colours with names being the names of pathways. If none is specified, a default colour scheme will automatically be chosen.

pathway

A character vector of length 1 specifying which pathway to plot, e.g. "clinical-mRNA".

nodeColours

A named vector of colours with names being "assay", "class1","class2". a default colour scheme will automatically be chosen.

classColours

A named vector of colours with names being "class1","class2", and "accuracy". a default colour scheme will automatically be chosen.