Convert Different Data Classes into DataFrame and Filter Features

Input data could be of matrix, MultiAssayExperiment, or DataFrame format and this function will prepare a DataFrame of features and a vector of outcomes and help to exclude nuisance features such as dates or unique sample identifiers from subsequent modelling.

Usage

# S4 method for matrix
prepareData(measurements, outcome, ...)

# S4 method for DataFrame
prepareData(
  measurements,
  outcome,
  useFeatures = "all",
  maxMissingProp = 0,
  topNvariance = NULL
)

# S4 method for MultiAssayExperiment
prepareData(measurements, outcomeColumns = NULL, useFeatures = "all", ...)

Arguments

measurements

Either a matrix, DataFrame or MultiAssayExperiment containing all of the data. For a matrix or DataFrame, the rows are samples, and the columns are features.

...

Variables not used by the matrix nor the MultiAssayExperiment method which are passed into and used by the DataFrame method.

outcome

Either a factor vector of classes, a Surv object, or a character string, or vector of such strings, containing column name(s) of column(s) containing either classes or time and event information about survival. If column names of survival information, time must be in first column and event status in the second.

useFeatures

If measurements is a MultiAssayExperiment, a two-column table of features to use. The first column must have assay names and the second column must have feature names found for that assay. "clinical" is also a valid assay name and refers to the clinical data table. "all" is a special keyword that means all features (passing any other filters) of that assay will be used for modelling. Otherwise, a character vector of feature names to use suffices.

maxMissingProp

Default: 0.0. A proportion less than 1 which is the maximum tolerated proportion of missingness for a feature to be retained for modelling.

topNvariance

Default: NULL. An integer number of most variable features to subset to.

outcomeColumns

If measurements is a MultiAssayExperiment, the names of the column (class) or columns (survival) in the table extracted by colData(data) that contain(s) the each individual's outcome to use for prediction.

Value

A list of length two. The first element is a DataFrame of features and the second element is the outcomes to use for modelling.

Author

Dario Strbenac